move queryFeaturedProductsCollection to fetchShopifyData.js, re-name FeaturedProduct to Product
