+++ src/Layout/Header.js
@@ -169,7 +169,6 @@ export class Header extends PureComponent {
   }
 
   render() {
-    console.log("props in header is: ", this.props)
   return (
     <>
       {this.createNavBar()}
