+++ src/SharedComponents/FeaturedProducts.js
@@ -5,9 +5,8 @@ import PropTypes from "prop-types";
 import { client } from "../plugins/shopify.js";
 import ComponentWrapper from "./ComponentWrapper";
 import StyledH2 from "./StyledH2";
-
-import FeaturedProduct from './FeaturedProduct';
-import { connect } from "react-redux";
+import Product from './Product';
+import { queryFeaturedProductsCollection } from '../state/fetchShopifyData';
 
 const ProductsContainer = styled.div`
   display: flex;
@@ -16,7 +15,6 @@ const ProductsContainer = styled.div`
   justify-content: space-evenly;
   position: relative;
 `;
-
 
 const ExploreShopLink = styled(Link)`
   display: block;
@@ -41,69 +39,17 @@ const ExploreShopLink = styled(Link)`
 const Products = ({ title, hasTopBottomBorders }) => {
   const [featuredProducts, setFeaturedProducts] = useState([]);
 
-  // create variable to use sortKey
-  const sortKey = client.graphQLClient.variable(
-    "sortKey",
-    "ProductCollectionSortKeys"
-  );
-  // query to get collection with handle === "featured-products"
-  const queryFeaturedProductsCollection = client.graphQLClient.query([sortKey], root => {
-    root.add(
-      "collectionByHandle",
-      { args: { handle: "featured-products" } },
-      collection => {
-        collection.add("id");
-        collection.addConnection(
-          "products",
-          { args: { sortKey: sortKey, first: 5} },
-          product => {
-            product.add("title");
-            product.add("descriptionHtml");
-            product.add("handle");
-            product.add("availableForSale");
-            product.addConnection(
-              "collections",
-              { args: { first: 2 } },
-              collection => {
-                collection.add("handle");
-              }
-            );
-            product.addConnection(
-              "metafields",
-              { args: { first: 2 } },
-              metafield => {
-                metafield.add("key");
-                metafield.add("value");
-              }
-            );
-            product.addConnection("images", { args: { first: 10 } }, image => {
-              image.add("id");
-              image.add("src");
-              image.add("altText");
-            });
-            product.addConnection(
-              "variants",
-              { args: { first: 1 } },
-              variant => {
-                variant.add("id");
-                variant.add("price");
-              }
-            );
-          }
-        );
-      }
-    );
-  });
-
-  useEffect(() => {
+  // TODO: figure out how to avoid re-fetching on route changes
+  if (featuredProducts.length === 0) {
     // make query with items sorted by MANUAL order
     client.graphQLClient
       .send(queryFeaturedProductsCollection, { sortKey: "MANUAL" })
       .then(({ model, data }) => {
         // store products in local state
         setFeaturedProducts(data.collectionByHandle.products.edges);
-      });
-    }, []);
+      })
+      .catch( error => console.log("Error fetching featured-products collection: ", error));
+  }
 
   return (
     <ComponentWrapper hasTopBottomBorders={hasTopBottomBorders}>
@@ -111,7 +57,7 @@ const Products = ({ title, hasTopBottomBorders }) => {
       <ProductsContainer>
         {featuredProducts
         .map(product => (
-          <FeaturedProduct key={product.id} product={product} />
+          <Product key={product.id} product={product} />
         ))}
       </ProductsContainer>
       <ExploreShopLink to="/shop">Explore the Shop</ExploreShopLink>
