+++ src/SharedComponents/Product.js
@@ -59,8 +59,6 @@ const createProduct = (product, clearHeroImg, updateQuantityButton) => {
     updateQuantityButton(1);
   }
 
-  console.log("what is product in FeaturedProduct?: ", product)
-
   return (
     <ProductLink to={`/product/${product.handle}`} onClick={clearHeroImgAndQuantityButton}>
       <ImageContainer>
