+++ src/state/fetchShopifyData.js
@@ -1,7 +1,62 @@
 import { fetchPending, fetchSuccess, fetchError } from './actions/cart';
 import { client } from "../plugins/shopify.js";
 
-// Here's the actions for fetching shopify data
+// FETCH "featured-products" COLLECTION
+
+  // create variable to use sortKey
+  const sortKey = client.graphQLClient.variable(
+    "sortKey",
+    "ProductCollectionSortKeys"
+  );
+  
+  // query to get collection with handle === "featured-products"
+  export const queryFeaturedProductsCollection = client.graphQLClient.query([sortKey], root => {
+    root.add(
+      "collectionByHandle",
+      { args: { handle: "featured-products" } },
+      collection => {
+        collection.add("id");
+        collection.addConnection(
+          "products",
+          { args: { sortKey: sortKey, first: 5} },
+          product => {
+            product.add("title");
+            product.add("descriptionHtml");
+            product.add("handle");
+            product.add("availableForSale");
+            product.addConnection(
+              "collections",
+              { args: { first: 2 } },
+              collection => {
+                collection.add("handle");
+              }
+            );
+            product.addConnection(
+              "metafields",
+              { args: { first: 2 } },
+              metafield => {
+                metafield.add("key");
+                metafield.add("value");
+              }
+            );
+            product.addConnection("images", { args: { first: 10 } }, image => {
+              image.add("id");
+              image.add("src");
+              image.add("altText");
+            });
+            product.addConnection(
+              "variants",
+              { args: { first: 1 } },
+              variant => {
+                variant.add("id");
+                variant.add("price");
+              }
+            );
+          }
+        );
+      }
+    );
+  });
 
 // create query for shopify articles, used in getArticles
 const articlesQuery = client.graphQLClient.query(root => {
